+++ src/app/components/modals/editar-espaco-modal/editar-espaco-modal.css
@@ -1,9 +1,9 @@
 .modal-geral {
-  width: 720px;
+  width: 45rem;
   border-radius: 8px;
   background: #fff;
   color: #222;
-  box-shadow: 0 6px 32px rgba(0,0,0,0.25);
+  box-shadow: 0 0.375rem 2rem rgba(0,0,0,0.25);
   overflow: hidden;
   margin: auto;
   font-family: Inter, sans-serif;
@@ -12,81 +12,85 @@
 .header-modal {
   background: var(--color-primary-300);
   color: #fff;
-  padding: 14px 18px;
+  padding: 0.875rem 1.125rem;
   display: flex;
   align-items: center;
   justify-content: center;
   position: relative;
 }
+
 .header-modal .title {
-  font-size: 18px;
+  font-size: 1.125rem;
   font-weight: 700;
 }
 
 .close-btn {
   position: absolute;
-  right: 14px;
-  top: 50%;
-  transform: translateY(-50%);
+  right: 0.875rem;
+  top: 0;
+  bottom: 0;
+  margin: auto 0;
+  height: fit-content;
   background: none;
   border: none;
-  font-size: 22px;
+  font-size: 1.375rem;
   color: #fff;
   cursor: pointer;
 }
 
 .body-modal {
-  padding: 20px;
+  padding: 1.25rem;
 }
 
 .form-grid {
   display: flex;
-  gap: 20px;
+  gap: 1.25rem;
 }
 
-.col-left, .col-right {
+.col-left,
+.col-right {
   flex: 1;
   display: flex;
   flex-direction: column;
 }
 
 .label-small {
-  font-size: 13px;
+  font-size: 0.8125rem;
   font-weight: 600;
-  margin-bottom: 4px;
+  margin-bottom: 0.25rem;
 }
 
 .input-small,
 .input-discipline {
-  padding: 6px 10px;
+  padding: 0.375rem 0.625rem;
   border: 1px solid #bfbfbf;
   border-radius: 4px;
   background: #f7f7f7;
 }
 
 .short-input {
-  width: 90px;
+  width: 5.625rem;
 }
 
 .disciplina-box {
-  margin-top: 4px;
+  margin-top: 0.25rem;
   border: 1px solid #c7c7c7;
   background: #f6f6f6;
-  padding: 8px;
+  padding: 0.5rem;
   border-radius: 4px;
-  min-height: 60px;
+  min-height: 3.75rem;
   display: flex;
-  gap: 6px;
+  gap: 0.375rem;
   flex-wrap: wrap;
 }
 
 .disciplina-item {
   background: #ececec;
-  padding: 4px 8px;
-  border-radius: 14px;
+  padding: 0.25rem 0.5rem;
+  border-radius: 0.875rem;
   display: flex;
   align-items: center;
-  gap: 6px;
+  gap: 0.375rem;
 }
 
 .chip-remove {
@@ -96,44 +100,44 @@
 }
 
 .add-disc-btn {
-    padding: 0.3rem 0.6rem;
+  padding: 0.3rem 0.6rem;
 }
 
 .indisponivel-row {
-  margin-top: 14px;
+  margin-top: 0.875rem;
   display: flex;
   align-items: center;
-  gap: 10px;
+  gap: 0.625rem;
 }
 
 .red-dot {
-  width: 18px;
-  height: 18px;
+  width: 1.125rem;
+  height: 1.125rem;
   border-radius: 50%;
   border: 3px solid white;
   background: var(--color-primary-300);
 }
 
 .footer-modal {
-  padding: 16px;
+  padding: 1rem;
   display: flex;
   justify-content: center;
-  gap: 20px;
+  gap: 1.25rem;
 }
 
 .btn-edit {
-  min-width: 180px;
+  min-width: 11.25rem;
   background: #e5e5e5;
   border: 1px solid #aaa;
-  padding: 10px;
+  padding: 0.625rem;
   font-weight: 700;
 }
 
 .btn-cancel {
-  min-width: 160px;
+  min-width: 10rem;
   background: var(--color-primary-300);
   border: 1px solid var(--color-primary-500);
   color: white;
-  padding: 10px;
+  padding: 0.625rem;
   font-weight: 700;
 }
