+++ src/app/components/modals/editar-espaco-modal/editar-espaco-modal.spec.ts
@@ -1,4 +1,5 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
+import type { ComponentFixture} from '@angular/core/testing';
+import { TestBed } from '@angular/core/testing';
 
 import { EditarEspacoModal } from './editar-espaco-modal';
