+++ src/app/components/secretaria/sala-form/sala-form.css
@@ -1,4 +1,4 @@
-section {
+form.form-group {
   max-width: 60rem;
   margin: 0 auto;
   padding: 2rem 1rem;
