+++ src/app/pages/space-management/space-management.ts
@@ -28,11 +28,11 @@ export class SpaceManagement implements OnInit {
 
   ngOnInit(): void {
     this.salaService.getSalas().subscribe({
-      next: (salas) => {
-        this.masterSalaList = salas;
+      next: (labs) => {
+        this.masterSalaList = labs;
         this.atualizarDataVisualizada();
       },
-      error: (err) => console.error('Erro ao carregar salas:', err)
+      error: (err) => console.error('Erro ao carregar laboratórios:', err)
     });
   }
